Fix the ivec usage for AABB

+++ src/graphics/AABB.cpp
@@ -1,6 +1,6 @@
 #include "AABB.hpp"
 
-AABB::AABB(const glm::ivec3& position, unsigned int size) :
+AABB::AABB(const glm::vec3& position, unsigned int size) :
 	position(position),
 	size(size)
 {
